# vals/OpenApiSpec.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "Skills API",
      "description": "API for learning various skills by fetching content from specified URLs.",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "https://xkonti-valtowncoderknowledge.web.val.run"
      }
    ],
    "paths": {
      "/learn/skill/{topic}": {
        "get": {
          "description": "Fetch content for a given skill topic.",
          "operationId": "learnSkill",
          "parameters": [
            {
              "name": "topic",
              "in": "path",
              "required": true,
              "description": "The name of the topic to learn.",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful retrieval of skill content.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "topic": {
                        "type": "string",
                        "description": "The name of the topic."
                      },
                      "skillContent": {
                        "type": "string",
                        "description": "The skill content in markdown format."
                      },
                      "extraTip": {
                        "type": "string",
                        "description": "Extra tip from the user."
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Topic not found."
            },
            "401": {
              "description": "Unauthorized - Missing or invalid API key."
            },
            "500": {
              "description": "Server Error - Issue with fetching the content."
            }
          },
          "security": [
            {
              "ApiKeyAuth": []
            }
          ]
        }
      }
    },
    "components": {
      "securitySchemes": {
        "ApiKeyAuth": {
          "type": "apiKey",
          "in": "header",
          "name": "x-api-key"
        }
      },
      "schemas": {}
    }
  }
  